+++ Assets/Dust/Scripts/Editor/Events/DuEventEditor.cs
@@ -5,14 +5,6 @@ namespace DustEngine.DustEditor
 {
     public abstract class DuEventEditor : DuEditor
     {
-        protected static DuEvent AddComponentByEventType(System.Type eventType)
-        {
-            if (Dust.IsNull(Selection.activeGameObject))
-                return null;
-
-            return Selection.activeGameObject.AddComponent(eventType) as DuEvent;
-        }
-
         public override void OnInspectorGUI()
         {
             // Hide base implementation