+++ Assets/Dust/CHANGELOG.md
@@ -4,6 +4,7 @@
 
 - DuTranslate renamed to DuMove
 - DuClampFactoryMachine: added flags to turn on/off PSR clamp for each or XYZ-axises 
+- On add Du*Event component it correct append to Undo stack
 
 
 ________________________________________________________________________________________________________________________